requirements.txt
PySocks
requests
stem
colorama
